+++ lib/ruby_aem/resources/package.rb
@@ -95,10 +95,15 @@ module RubyAem
       #
       # @param file_path the directory where the package file to be uploaded is
       # @param force if true, then overwrite if the package already exists
+      # @param opts optional parameters:
+      # - force: if false then a package file will not be uploaded when the package already exists with the same group, name, and version, default is true (will overwrite existing package file)
       # @return RubyAem::Result
-      def upload(file_path, force)
+      def upload(file_path,
+        opts = {
+          force: true
+        })
         @call_params[:file_path] = file_path
-        @call_params[:force] = force
+        @call_params = @call_params.merge(opts)
         @client.call(self.class, __callee__.to_s, @call_params)
       end